update mutation code (#1453)
